+++ formulae/terms/call.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from pandas.api.types import is_categorical_dtype, is_numeric_dtype, is_string_dtype
 
-from formulae.transforms import TRANSFORMS
+from formulae.transforms import TRANSFORMS, Prop, Offset
 from formulae.terms.call_utils import CallVarsExtractor
 
 
@@ -99,9 +99,12 @@ class Call:
 
         if is_numeric_dtype(x):
             self._type = "numeric"
-        # Improve this condition, the dictionary may cause problems!
-        elif is_string_dtype(x) or is_categorical_dtype(x) or isinstance(x, dict):
+        elif is_string_dtype(x) or is_categorical_dtype(x):
             self._type = "categoric"
+        elif isinstance(x, Prop):
+            self._type = "prop"
+        elif isinstance(x, Offset):
+            self._type = "offset"
         else:
             raise ValueError(f"Call result is of an unrecognized type ({type(x)}).")
         self._intermediate_data = x
@@ -124,12 +127,16 @@ class Call:
         try:
             if self._type is None:
                 raise ValueError("Call result type is not set.")
-            if self._type not in ["numeric", "categoric"]:
+            if self._type not in ["numeric", "categoric", "prop", "offset"]:
                 raise ValueError(f"Call result is of an unrecognized type ({self._type}).")
             if self._type == "numeric":
                 self.data = self._eval_numeric(self._intermediate_data)
-            else:
+            elif self._type == "categoric":
                 self.data = self._eval_categoric(self._intermediate_data, encoding)
+            elif self._type == "prop":
+                self.data = self._eval_prop(self._intermediate_data)
+            elif self._type == "offset":
+                self.data = self._eval_offset(self._intermediate_data)
         except:
             print("Unexpected error while trying to evaluate a Call:", sys.exc_info()[0])
             raise
@@ -153,11 +160,12 @@ class Call:
             evaluation, and the latter is equal to ``"numeric"``.
         """
         if isinstance(x, np.ndarray):
-            value = np.atleast_2d(x)
-            if x.shape[0] == 1 and x.shape[1] > 1:
-                value = value.T
+            value = x.flatten()
+            if value.ndim > 1:
+                raise ValueError(f"The result of {self.name} is not 1-dimensional.")
+            value = value[:, np.newaxis]
         elif isinstance(x, pd.Series):
-            value = np.atleast_2d(x.to_numpy()).T
+            value = x.to_numpy()[:, np.newaxis]
         else:
             raise ValueError(f"Call result is of an unrecognized type ({type(x)}).")
         return {"value": value, "type": "numeric"}
@@ -220,6 +228,16 @@ class Call:
             "encoding": encoding,
         }
 
+    def _eval_prop(self, prop):
+        if not self.is_response:
+            raise ValueError("'prop()' can only be used in the context of a response term.")
+        return {"value": prop.eval(), "type": "prop"}
+
+    def _eval_offset(self, offset):
+        if self.is_response:
+            raise ValueError("'offset() cannot be used in the context of a response term.")
+        return {"value": offset.eval(), "type": "offset"}
+
     def eval_new_data(self, data_mask):
         """Evaluates the function call with new data.
 
@@ -261,17 +279,15 @@ class Call:
             number of dummy variables used in the numeric representation of the categorical
             variable.
         """
-        if self.is_response:
-            return np.atleast_2d(np.where(x == self.data["reference"], 1, 0)).T
+
+        # Raise error if passing a level that was not observed.
+        new_data_levels = pd.Categorical(x).dtype.categories.tolist()
+        if set(new_data_levels).issubset(set(self.data["levels"])):
+            series = pd.Categorical(x, categories=self.data["levels"])
+            drop_first = self.data["encoding"] == "reduced"
+            return pd.get_dummies(series, drop_first=drop_first).to_numpy()
         else:
-            # Raise error if passing a level that was not observed.
-            new_data_levels = pd.Categorical(x).dtype.categories.tolist()
-            if set(new_data_levels).issubset(set(self.data["levels"])):
-                series = pd.Categorical(x, categories=self.data["levels"])
-                drop_first = self.data["encoding"] == "reduced"
-                return pd.get_dummies(series, drop_first=drop_first).to_numpy()
-            else:
-                raise ValueError(
-                    f"At least one of the levels for '{self.name}' in the new data was "
-                    "not present in the original data set."
-                )
+            raise ValueError(
+                f"At least one of the levels for '{self.name}' in the new data was "
+                "not present in the original data set."
+            )