+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## v0.1.x Unreleased
+
+### New features
+
+- Added `prop()` function to handle response terms that a proportion computed from two variables (#40)
+- Categorical responses can contain more than one level (#40)
+- Added `binary()` function to convert numeric/categorical variables into 0-1 variables (#41)
+- Modified `C()` to make it simpler to obtain categorical variables and manipulate their levels (#41)
+- Added `offset()` (#42)
+
 ## v0.1.3
 
 - Temporary release